fix: rename filename to remove duplicate `redis`
